requirements.txt
requests
google-auth-oauthlib
google-api-python-client
youtube-transcript-api
